feat(enrichment): add grouping logic for Aal miRNA UniProt IDs to facilitate Venny analysis
Added code to `01-aal-save-uniprot-ids-txt.R` to group the 24 Aal miRNA
names into 4 sets of 6 for input into the Venny web tool. This includes:
- Sorting miRNA names using `gtools::mixedsort` for natural order (e.g.,
miR-2a before miR-10)
- Splitting the sorted list into 4 groups
- Printing the groups to console for manual copy/paste into Venny

Also added `library(gtools)` dependency. This package helped sorting
complex strings with letters and numbers such as `aal-miR-2a-5p` with
the miRNA with the smallest number at the start of the list to the
biggest number at the end of the list.