refactor(enrichment): comment out filtering of NA categories and update comments
Modified `01-fuse-enrichment-results.R` to comment out the code block responsible for filtering rows with NA values in the `category_of_interest` column. This change affects the processing of per-miRNA, venny, and all STRINGdb enrichment results. Updated comments to indicate that this filtering step is now optional and clarified the purpose of the remaining export section which deals with the unfiltered (important) data. The script continues to save the full (unfiltered) fused data frames.